+++ weave/tests/test_wb_end_to_end.py
@@ -1,5 +1,10 @@
+import time
 import weave
 import wandb
+from weave import weave_internal
+
+from weave.ecosystem.wandb.panel_time_series import TimeSeries
+
 
 # Example of end to end integration test
 def test_run_logging(user_by_api_key_in_env):
@@ -55,3 +60,61 @@ def _test_basic_publish(user_fixture):
         == f"wandb-artifact:///{user_fixture.username}/weave/list:0cdf3358dc939f961ca9/obj"
     )
     assert weave.ref_base.Ref.from_str(uri).get() == [1, 2, 3]
+
+
+def test_compile_through_execution(user_by_api_key_in_env):
+    """
+    This test demonstrates successful execution when there is an explicit
+    const function instead of a direct node (resulting in an intermediate execution op)
+    """
+    run = wandb.init(project="project_exists")
+    for i in range(10):
+        run.log({"val": i, "cat": i % 2})
+    run.finish()
+
+    history_node = weave.ops.project(run.entity, run.project).run(run.id).history2()
+    pick = weave_internal.const(history_node).pick("val")
+    res = weave.use(pick)
+    assert res.to_pylist_notags() == list(range(10))
+
+
+def test_compile_through_function_call(user_by_api_key_in_env):
+    """
+    This test demonstrates successful execution when there is an explicit
+    function-__call__ in the graph)
+    """
+
+    run = wandb.init(project="project_exists")
+    for i in range(10):
+        run.log({"val": i, "cat": i % 2})
+    run.finish()
+
+    const_node = weave_internal.const(
+        weave_internal.define_fn(
+            {"entity_name": types.String()},
+            lambda entity_name: (
+                weave.ops.project(entity_name, run.project).run(run.id).history2()
+            ),
+        )
+    )(run.entity)
+    pick = const_node.pick("val")
+    res = weave.use(pick)
+    assert res.to_pylist_notags() == list(range(10))
+
+
+def test_panel_timeseries(user_by_api_key_in_env):
+    run = wandb.init(project="project_exists")
+    for i in range(10):
+        time.sleep(0.2)
+        run.log({"val": i, "cat": str(i % 2)})
+    run.finish()
+
+    history_node = weave.ops.project(run.entity, run.project).run(run.id).history2()
+    panel = TimeSeries(history_node)
+    init_config_node = panel.initialize()
+    init_config = weave.use(init_config_node)
+    panel.config = init_config
+    render_node = panel.render()
+    res = weave.use(render_node)
+    # What to assert here? Should we be getting .contents?
+    assert res != None
